+++ src/utils/citation_utils.tsx
@@ -320,7 +320,6 @@ export function parseCitationsInMarkdown(md: string): CitationRef[] {
 
 export interface SpanStyles {
     citationColorInPdf: string;
-    superScriptColorInPdf: string;
 }
 
 /**
@@ -503,8 +502,6 @@ function processInlineReferences(
  */
 const DEFAULT_CONTAINER_STYLE = 'cursor: default;';
 const DEFAULT_CITATION_STYLE = 'text-decoration: none; cursor: pointer;';
-const DEFAULT_SUPERSCRIPT_STYLE = 'font-size: 0.7em; vertical-align: super; margin-left: 1px; text-decoration: none; cursor: pointer;';
-
 
 /**
  * Generates span tags for citations
@@ -518,10 +515,7 @@ export function generateCitationSpans(
 ): string {
     const spans = citations.map((citation, index) => {
         const { local, crossFile } = splitFileCitation(citation, fileDelimiter);
-
-
         const citationColorInPdf = spanStyles.citationColorInPdf || '#000000';
-        const superScriptColorInPdf = spanStyles.superScriptColorInPdf || '#000000';
 
         // Combine default styles with color
         const containerStyle = escapeString(
@@ -533,17 +527,11 @@ export function generateCitationSpans(
             DEFAULT_CITATION_STYLE + ` color: ${citationColorInPdf};`,
             "\""
         );
-
-        const superscriptStyle = escapeString(
-            DEFAULT_SUPERSCRIPT_STYLE + ` color: ${superScriptColorInPdf};`,
-            "\""
-        );
-
         let result = `<span style="${containerStyle}">` + citationFormat.replace('#', `<span style="${citationStyle}">${local}</span>`);
-        if (crossFile) {
-            result += `<sup style="${superscriptStyle}">${'[' + crossFile + ']'}</sup>`;
-        }
         result += '</span>';
+        if (crossFile) {
+            result += `${'[^' + crossFile + ']'}`;
+        } 
 
         return result;
     });